Work in progress

+++ openglrenderer/graphview.h
@@ -1,5 +1,5 @@
 #include <stdio.h>
 
 void graphview_init(int width, int height, int fullscreen);
-int graphview_update(double *dt);
+int graphview_update(float *dt);
 void graphview_shutdown();
